fix: cache molecule instances instead of svg
See #24

+++ src/utils/molecule.ts
@@ -1,4 +1,5 @@
 import { RDKitModule } from '@rdkit/rdkit';
+import { getJSMolFromCache, storeJSMolInCache } from './caching';
 
 export const get_molecule_details = (
   smiles: string,
@@ -8,7 +9,6 @@ export const get_molecule_details = (
   if (!mol) return null;
 
   const details = JSON.parse(mol.get_descriptors());
-  mol.delete();
   return {
     numAtoms: details.NumHeavyAtoms,
     numRings: details.NumRings,
@@ -20,7 +20,6 @@ export const is_valid_smiles = (smiles: string, RDKit: RDKitModule): boolean => 
   const mol = get_molecule(smiles, RDKit);
   if (!mol) return false;
   const isValid = mol.is_valid();
-  mol.delete();
   return isValid;
 };
 
@@ -29,7 +28,6 @@ export const get_canonical_form_for_structure = (structure: string, RDKit: RDKit
   const mol = get_molecule(structure, RDKit);
   if (!mol) return null;
   const cannonicalForm = mol.get_smarts();
-  mol.delete();
   return cannonicalForm;
 };
 
@@ -37,16 +35,20 @@ const get_canonical_smiles = (smiles: string, RDKit: RDKitModule): string | null
   const mol = get_molecule(smiles, RDKit);
   if (!mol) return null;
   const cannonicalSmiles = mol.get_smiles();
-  mol.delete();
   return cannonicalSmiles;
 };
 
 export const get_molecule = (smiles: string, RDKit: RDKitModule) => {
+  const cachedMolecule = getJSMolFromCache(smiles);
+  if (cachedMolecule) return cachedMolecule;
+  if (!smiles) return null;
   if (!RDKit) return null;
 
   const tempMolecule = RDKit.get_mol(smiles);
   const mdlWithCoords = tempMolecule.get_new_coords(true);
   tempMolecule.delete();
 
-  return RDKit.get_mol(mdlWithCoords);
+  const mol = RDKit.get_mol(mdlWithCoords);
+  storeJSMolInCache(smiles, mol);
+  return mol;
 };
